ci: Implement smoke tests for pkg/cmd, fix examples
